Finished foodCard grabs data from database
